backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
django-filter==23.2 